Moved some utility functions to utils.
